requirements.txt
streamlit
pandas
python-docx
pypandoc
reportlab
PyPDF2
openpyxl
pydrive